+++ components/src/lib/AlertModal/AlertModal.styles.ts
